The lowerDomain incorrectly was initialized to the byteValue rather than the doubleValue.  This was problematic for negative lower domains.
git-svn-id: http://chianti.ucsd.edu/svn@21822 0ecc0d97-ab19-0410-9704-bfe1a75892f5

+++ cytoscape/trunk/application/src/main/java/cytoscape/visual/mappings/continuous/ContinuousRangeCalculator.java
@@ -158,7 +158,7 @@ public class ContinuousRangeCalculator {
     private Object getRangeValue(int index, Number domainValue) {
         //  Get Lower Domain and Range
         ContinuousMappingPoint lowerBound = points.get(index - 1);
-        Number lowerDomain = lowerBound.getValue().byteValue();
+        Number lowerDomain = lowerBound.getValue().doubleValue();
         BoundaryRangeValues lv = lowerBound.getRange();
         Object lowerRange = lv.greaterValue;
